Major refactoring to scripting layer.
The most important change is avoidance of light userdata in favor of rlua's new userdata facilities. This avoids a lot of crash-prone pointer magic, and makes for a cleaner scripting API.

Unfortunately, with the change it is no longer possible to use references to transactions - the transactions need to be owned by the script executor. As a consequence, script execution is no longer available in the bulk API.